@@ -1,11 +1,11 @@
 name: chips_input
-description: A Flutter Widget to create an input with multiple chips, that can be selected from suggestions. 
+description: A Flutter Widget to create an input with multiple chips, that can be selected from suggestions.
 version: 0.1.0
 homepage: https://github.com/zsoerenm/chips_input
 
 environment:
-  sdk: ">=2.12.0 <3.0.0"
-  flutter: ">=2.0.0"
+  sdk: '>=3.0.0 <4.0.0'
+  flutter: ">=3.0.0"
 
 dependencies:
   flutter:
